src/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.9

